requirements.txt
aiodns==3.0.0
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-socks==0.4.1
aiomultiprocess==0.9.0
aioresponses==0.7.3
aiosignal==1.2.0
aiosqlite==0.17.0
altgraph==0.17.2
anyio==3.6.1
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backoff==2.1.2
beautifulsoup4==4.11.1
cachetools==5.2.0
cchardet==2.1.7
censys==2.1.6
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
chromedriver-autoinstaller==0.3.1
click==8.1.3
click-plugins==1.1.1
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.2
cycler==0.11.0
dataclasses==0.6
dataclasses-json==0.5.7
decorator==5.1.1
distlib==0.3.4
dnspython==2.2.1
elastic-transport==8.1.2
elasticsearch==5.5.3
fake-useragent==0.1.11
fastapi==0.78.0
filelock==3.7.1
flake8==4.0.1
fonttools==4.34.4
frozenlist==1.3.0
gcloud==0.18.3
geckodriver-autoinstaller==0.1.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
googletransx==2.4.2
gspread==5.4.0
gspread-pandas==3.2.2
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httplib2shim==0.0.3
httpretty==1.1.4
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
kiwisolver==1.4.3
limiter==0.1.2
limits==1.6
lxml==4.9.0
macholib==1.16
marshmallow==3.16.0
marshmallow-enum==1.5.1
matplotlib==3.5.2
maxminddb==2.2.0
mccabe==0.6.1
more-itertools==8.13.0
multidict==6.0.2
mypy==0.961
mypy-extensions==0.4.3
nassl==4.0.2
netaddr==0.8.0
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
ppdeep==20200505
protobuf==4.21.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.2.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyee==8.2.2
pyflakes==2.4.0
Pygments==2.12.0
pygsheets==2.0.5
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pymongo==4.1.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
PySocks==1.7.1
pytest==7.1.2
pytest-asyncio==0.18.3
python-certifi-win32==1.6
python-dateutil==2.8.2
python-dotenv==0.20.0
python-socks==2.0.3
python-Wappalyzer==0.3.1
pytz==2022.1
PyYAML==6.0
redis==3.5.3
requests==2.28.0
requests-futures==1.0.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
responses==0.21.0
retrying==1.3.3
rfc3986==1.5.0
rich==12.4.4
rsa==4.8
schedule==1.1.0
selenium==4.2.0
setuptools-scm==7.0.2
shodan==1.27.0
six==1.16.0
slowapi==0.1.5
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
splunk-sdk==1.7.0
sslyze==5.0.5
starlette==0.19.1
Sublist3r==1.0
termcolor==1.1.0
tls-parser==2.0.0
token-bucket==0.2.0
tomli==2.0.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
twitter==1.19.3
types-certifi==2021.10.8.3
types-chardet==4.0.4
types-PyYAML==6.0.8
types-requests==2.27.30
types-ujson==5.3.0
types-urllib3==1.26.15
typing-inspect==0.7.1
typing_extensions==4.2.0
ujson==5.3.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.18.1
uvloop==0.16.0
virtualenv==20.14.1
websockets==10.3
whois==0.9.16
wrapt==1.14.1
wsproto==1.1.0
XlsxWriter==3.0.3
yara-python==4.2.0
yarl==1.7.2
zipp==3.8.0
